requirements.txt
pytest==8.2.2
webdriver-manager==4.0.2
selenium==4.22.0
requests==2.32.3
allure-pytest==2.13.5
allure-python-commons==2.13.5